chore(test): moving the test in the appropriate file
