+++ test/membrane/symbols.spec.js
@@ -43,4 +43,30 @@ describe('Secure Membrane', () => {
             expect(Symbol.for('symbol-with-key').constructor === Symbol).toBeTrue();
         `);
     });
+
+    it('blue Symbol class properties are inherited in red environments', () => {
+        const symbol = Symbol.for('method');
+        
+        let successful = false;
+        
+        class Base {
+            [symbol]() {
+                successful = true;
+            }
+        }
+
+        const secureEval = createSecureEnvironment(undefined, { Base, symbol });
+        secureEval(`
+            class Bar extends Base {
+                constructor() {
+                    super();
+                    this.barClass = 'bar';
+                    this[symbol]();
+                }
+            }
+            new Bar();
+        `);
+
+        expect(successful).toBe(true);
+    });
 });
